+++ lib/vy_date.dart
@@ -1,6 +1,5 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
+/// Copyright © 2018 Giorgio Franceschetti. All rights reserved.
+
 library vy_date;
 
 export 'src/date.dart';